// generated/src/app/api-models/pmb-admin-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbAdminWork {
  seqGpbilId: number;
  seqAdminId: number;
  customerType: string;
  customerId: string;
  groupId: string;
  invoiceNo: number;
  billDetailFromDate: Date;
  billDetailThruDate: Date;
  planCode: string;
  premiumAmt: number;
  seqParentId: number;
}